refactor(express-api): standardise error transformations
